+++ src/runtime/RuntimeType.ts
@@ -2,12 +2,13 @@ import { TypeDeep } from "../base/type-deep.js";
 import {
   LikeBuffer_t,
   IBufferLike,
+  INestedBufferLike,
   IDecodeOptions,
   IEncodeOptions,
 } from "../interfaces.js";
 
 export default abstract class RuntimeType<D, E>
-  extends TypeDeep<IBufferLike<D[], E[]>>
+  extends TypeDeep<INestedBufferLike<D[], E[]>>
   implements IBufferLike<D, E>
 {
   protected _byteLength: number = 0;